+++ packages/sdk/src/sdk-blockchains/ethereum/common/index.ts
@@ -1,12 +1,23 @@
 import type { Address, UnionAddress, Word } from "@rarible/types"
-import { toAddress, toContractAddress, toItemId, toOrderId, toUnionAddress } from "@rarible/types"
+import {
+	toAddress,
+	toBigNumber,
+	toBinary,
+	toContractAddress,
+	toItemId,
+	toOrderId,
+	toUnionAddress,
+} from "@rarible/types"
 import { isRealBlockchainSpecified } from "@rarible/types/build/blockchains"
 import type { AssetType, ItemId, OrderId } from "@rarible/api-client"
 import { Blockchain } from "@rarible/api-client"
 import type { UnionPart } from "packages/sdk/src/types/order/common"
-import type { Erc20AssetType, EthAssetType, Part } from "@rarible/ethereum-api-client"
+import type { Part } from "@rarible/ethereum-api-client"
 import type { ContractAddress } from "@rarible/types/build/contract-address"
 import type { EthereumNetwork } from "@rarible/protocol-ethereum-sdk/build/types"
+import { toBn } from "@rarible/utils/build/bn"
+import type { AssetType as EthereumAssetType } from "@rarible/ethereum-api-client/build/models/AssetType"
+import type { Order } from "@rarible/ethereum-api-client/build/models"
 import type { CurrencyType, RequestCurrency } from "../../../common/domain"
 
 export type EVMBlockchain = Blockchain.ETHEREUM | Blockchain.POLYGON
@@ -31,7 +42,7 @@ export function getEthTakeAssetType(currency: RequestCurrency) {
 	}
 }
 
-export function convertToEthereumAssetType(assetType: AssetType): EthAssetType | Erc20AssetType {
+export function convertToEthereumAssetType(assetType: AssetType): EthereumAssetType {
 	switch (assetType["@type"]) {
 		case "ETH": {
 			return { assetClass: "ETH" }
@@ -39,6 +50,68 @@ export function convertToEthereumAssetType(assetType: AssetType): EthAssetType |
 		case "ERC20": {
 			return {
 				assetClass: "ERC20",
+				contract: convertToEthereumAddress(assetType.contract),
+			}
+		}
+		case "ERC721": {
+			return {
+				assetClass: "ERC721",
+				contract: convertToEthereumAddress(assetType.contract),
+				tokenId: assetType.tokenId,
+			}
+		}
+		case "ERC721_Lazy": {
+			return {
+				assetClass: "ERC721_LAZY",
+				contract: convertToEthereumAddress(assetType.contract),
+				tokenId: assetType.tokenId,
+				uri: assetType.uri,
+				creators: assetType.creators.map(c => ({
+					account: convertToEthereumAddress(c.account),
+					value: toBn(c.value).toNumber(),
+				})),
+				royalties: assetType.royalties.map(r => ({
+					account: convertToEthereumAddress(r.account),
+					value: toBn(r.value).toNumber(),
+				})),
+				signatures: assetType.signatures.map(str => toBinary(str)),
+			}
+		}
+		case "ERC1155": {
+			return {
+				assetClass: "ERC1155",
+				contract: convertToEthereumAddress(assetType.contract),
+				tokenId: assetType.tokenId,
+			}
+		}
+		case "ERC1155_Lazy": {
+			return {
+				assetClass: "ERC1155_LAZY",
+				contract: convertToEthereumAddress(assetType.contract),
+				tokenId: assetType.tokenId,
+				uri: assetType.uri,
+				supply: assetType.supply !== undefined ? toBigNumber(assetType.supply): toBigNumber("1"),
+				creators: assetType.creators.map(c => ({
+					account: convertToEthereumAddress(c.account),
+					value: toBn(c.value).toNumber(),
+				})),
+				royalties: assetType.royalties.map(r => ({
+					account: convertToEthereumAddress(r.account),
+					value: toBn(r.value).toNumber(),
+				})),
+				signatures: assetType.signatures.map(str => toBinary(str)),
+			}
+		}
+		case "CRYPTO_PUNKS": {
+			return {
+				assetClass: "CRYPTO_PUNKS",
+				contract: convertToEthereumAddress(assetType.contract),
+				tokenId: assetType.tokenId,
+			}
+		}
+		case "GEN_ART": {
+			return {
+				assetClass: "GEN_ART",
 				contract: convertToEthereumAddress(assetType.contract),
 			}
 		}
@@ -54,6 +127,19 @@ export function toEthereumParts(parts: UnionPart[] | undefined): Part[] {
 	})) || []
 }
 
+export function getOriginFeesSum(originFees: Array<Part>): number {
+	return originFees.reduce((acc, fee) => fee.value, 0)
+}
+
+export function getOrderFeesSum(order: Order): number {
+	switch (order.data.dataType) {
+		case "LEGACY": return order.data.fee
+		case "RARIBLE_V2_DATA_V1": return getOriginFeesSum(order.data.originFees)
+		case "RARIBLE_V2_DATA_V2": return getOriginFeesSum(order.data.originFees)
+		default: throw new Error("Unexpected order dataType")
+	}
+}
+
 export function getEVMBlockchain(network: EthereumNetwork): EVMBlockchain {
 	switch (network) {
 		case "e2e":
@@ -65,6 +151,8 @@ export function getEVMBlockchain(network: EthereumNetwork): EVMBlockchain {
 		case "mainnet":
 			return Blockchain.ETHEREUM
 		case "mumbai":
+			return Blockchain.POLYGON
+		case "mumbai-dev":
 			return Blockchain.POLYGON
 		case "polygon":
 			return Blockchain.POLYGON
@@ -111,7 +199,7 @@ export function convertEthereumContractAddress(address: string, blockchain: EVMB
 	return toContractAddress(`${blockchain}:${address}`)
 }
 
-export function convertEthereumUnionAddress(address: string, blockchain: EVMBlockchain): UnionAddress {
+export function convertEthereumToUnionAddress(address: string, blockchain: EVMBlockchain): UnionAddress {
 	return toUnionAddress(`${blockchain}:${address}`)
 }