+++ packages/sdk/src/sdk-blockchains/flow/bid.ts
@@ -4,11 +4,15 @@ import { Action } from "@rarible/action"
 import { toFlowItemId } from "@rarible/flow-sdk/build/common/item"
 import { toBigNumber } from "@rarible/types/build/big-number"
 import { Blockchain } from "@rarible/api-client"
-import type { IBlockchainTransaction } from "@rarible/sdk-transaction"
 import { OriginFeeSupport, PayoutsSupport } from "../../types/order/fill/domain"
 import type * as OrderCommon from "../../types/order/common"
 import type { CurrencyType } from "../../common/domain"
-import type { GetConvertableValueResult, PrepareBidRequest, PrepareBidResponse } from "../../types/order/bid/domain"
+import type {
+	GetConvertableValueResult,
+	PrepareBidRequest,
+	PrepareBidResponse,
+	PrepareBidUpdateResponse,
+} from "../../types/order/bid/domain"
 import { convertFlowContractAddress, convertFlowOrderId, getFungibleTokenName, toFlowParts } from "./common/converters"
 import { getFlowBaseFee } from "./common/get-flow-base-fee"
 
@@ -25,10 +29,6 @@ export class FlowBid {
 
 	private async getConvertableValue(): Promise<GetConvertableValueResult> {
 		return undefined
-	}
-
-	private async convert(): Promise<IBlockchainTransaction> {
-		throw new Error("Convert operation is not supported")
 	}
 
 	async bid(prepare: PrepareBidRequest): Promise<PrepareBidResponse> {
@@ -70,14 +70,13 @@ export class FlowBid {
 			maxAmount: toBigNumber("1"),
 			baseFee: getFlowBaseFee(this.sdk),
 			getConvertableValue: this.getConvertableValue,
-			convert: this.convert,
 			submit: bidAction,
 		}
 	}
 
 	async update(
 		prepareRequest: OrderCommon.PrepareOrderUpdateRequest,
-	): Promise<OrderCommon.PrepareOrderUpdateResponse> {
+	): Promise<PrepareBidUpdateResponse> {
 		if (!prepareRequest.orderId) {
 			throw new Error("OrderId has not been specified")
 		}
@@ -108,6 +107,7 @@ export class FlowBid {
 			payoutsSupport: PayoutsSupport.NONE,
 			supportedCurrencies: FlowBid.supportedCurrencies,
 			baseFee: getFlowBaseFee(this.sdk),
+			getConvertableValue: this.getConvertableValue,
 			submit: bidUpdateAction,
 		}
 	}