+++ packages/sdk/src/sdk-blockchains/tezos/transfer.test.ts
@@ -1,13 +1,14 @@
 import { toContractAddress, toUnionAddress } from "@rarible/types"
 import { createRaribleSdk } from "../../index"
 import { MintType } from "../../types/nft/mint/domain"
+import { LogsLevel } from "../../domain"
 import { createTestWallet } from "./test/test-wallet"
 import { awaitForOwnership } from "./test/await-for-ownership"
 import { awaitForItemSupply } from "./test/await-for-item-supply"
 
 describe("transfer test", () => {
 	const wallet = createTestWallet("edsk3UUamwmemNBJgDvS8jXCgKsvjL2NoTwYRFpGSRPut4Hmfs6dG8")
-	const sdk = createRaribleSdk(wallet, "dev")
+	const sdk = createRaribleSdk(wallet, "dev", { logs: LogsLevel.DISABLED })
 
 	const receipent = "tz1VXxRfyFHoPXBVUrWY5tsa1oWevrgChhSg"
 	let nftContract: string = "KT1Ctz9vuC6uxsBPD4GbdbPaJvZogWhE9SLu"
@@ -66,7 +67,6 @@ describe("transfer test", () => {
 		})
 		await result.wait()
 
-		console.log("mintResult.itemId", mintResult.itemId, receipent)
 		await awaitForOwnership(sdk, mintResult.itemId, receipent)
 
 	}, 1500000)