+++ packages/sdk/src/sdk-blockchains/union/index.ts
@@ -17,6 +17,8 @@ import type { IDeploy } from "../../types/nft/deploy/domain"
 import type { CanTransferResult, IRestrictionSdk } from "../../types/nft/restriction/domain"
 import type { PreprocessMetaRequest, PreprocessMetaResponse } from "../../types/nft/mint/preprocess-meta"
 import type { PrepareBidRequest, PrepareBidResponse } from "../../types/order/bid/domain"
+import { Middlewarer } from "../../common/middleware/middleware"
+import type { PrepareBidUpdateResponse } from "../../types/order/bid/domain"
 
 export function createUnionSdk(
 	ethereum: IRaribleInternalSdk,
@@ -67,7 +69,7 @@ class UnionOrderSdk implements IOrderInternalSdk {
 		return this.instances[extractBlockchain(getBidEntity(request))].bid(request)
 	}
 
-	bidUpdate(request: OrderCommon.PrepareOrderUpdateRequest): Promise<OrderCommon.PrepareOrderUpdateResponse> {
+	bidUpdate(request: OrderCommon.PrepareOrderUpdateRequest): Promise<PrepareBidUpdateResponse> {
 		return this.instances[extractBlockchain(request.orderId)].bidUpdate(request)
 	}
 
@@ -114,7 +116,7 @@ class UnionNftSdk implements Omit<INftSdk, "mintAndSell"> {
 		this.burn = this.burn.bind(this)
 		this.mint = this.mint.bind(this)
 		this.transfer = this.transfer.bind(this)
-		this.preprocessMeta = this.preprocessMeta.bind(this)
+		this.preprocessMeta = Middlewarer.skipMiddleware(this.preprocessMeta.bind(this))
 		this.generateTokenId = this.generateTokenId.bind(this)
 	}