+++ packages/sdk/src/types/order/bid/domain.ts
@@ -1,25 +1,35 @@
 import type { AssetType } from "@rarible/api-client"
 import type { BigNumberValue } from "@rarible/utils"
-import type { IBlockchainTransaction } from "@rarible/sdk-transaction"
-import type { ContractAddress, UnionAddress } from "@rarible/types"
+import type { ContractAddress } from "@rarible/types"
 import type {
 	PrepareOrderRequest,
 	PrepareOrderResponse,
 	PrepareOrderUpdateRequest,
-	PrepareOrderUpdateResponse,
+	PrepareOrderUpdateResponse, UnionPart,
 } from "../common"
 
-export type PrepareBidResponse = PrepareOrderResponse & {
-	getConvertableValue(
-		assetType: AssetType, value: BigNumberValue, walletAddress: UnionAddress
-	): Promise<GetConvertableValueResult>
-	convert(from: AssetType, to: AssetType, value: BigNumberValue): Promise<IBlockchainTransaction>
+export type PrepareBidResponse = PrepareOrderResponse & GetConvertableValueFunction
+
+export type PrepareBidUpdateResponse = PrepareOrderUpdateResponse & GetConvertableValueFunction
+
+export type GetConvertableValueFunction = {
+	getConvertableValue(request: GetConvertableValueRequest): Promise<GetConvertableValueResult>
 }
 
-export type PrepareBidRequest = PrepareOrderRequest | { collectionId: ContractAddress }
-
-export type IBid = (request: PrepareBidRequest) => Promise<PrepareBidResponse>
-export type IBidUpdate = (request: PrepareOrderUpdateRequest) => Promise<PrepareOrderUpdateResponse>
-
+export type GetConvertableValueRequest = {
+	assetType: AssetType
+	price: BigNumberValue
+	amount: number
+	originFees: UnionPart[]
+}
 export type GetConvertableValueResult =
   { type: "insufficient" | "convertable", currency: AssetType; value: BigNumberValue } | undefined
+
+
+export type IBid = (request: PrepareBidRequest) => Promise<PrepareBidResponse>
+export type IBidUpdate = (request: PrepareOrderUpdateRequest) => Promise<PrepareBidUpdateResponse>
+export type PrepareBidRequest = PrepareOrderRequest | { collectionId: ContractAddress }
+
+export type ConvertCurrencyRequest = {
+	price: BigNumberValue
+}